+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.36.0] - 2021-01-02 Hello Hierarchy!
 
 ### Added